# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Arkeo Racing API
  description: API for the Arkeo Racing game - a horse racing simulation powered by blockchain service provider metrics.
  version: 0.1.0
  contact:
    name: Arkeo Racing
servers:
  - url: /api
    description: API base path
...
paths:
...
  /me/wallet:
    get:
      tags: [Users]
      summary: Get wallet info
      description: Returns wallet address and chain configuration
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Wallet and chain info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Users]
      summary: Connect wallet
      description: Links a wallet address to the user session
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [walletAddress]
              properties:
                walletAddress:
                  type: string
                  description: Arkeo wallet address (bech32)
                  example: arkeo1abc123...
                nickname:
                  type: string
                  description: Display nickname (3-24 chars, unique)
                  example: RaceFan42
      responses:
        '200':
          description: Wallet connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /me/nickname:
    post:
      tags: [Users]
      summary: Update nickname
      description: Sets or updates user display nickname
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nickname]
              properties:
                nickname:
                  type: string
                  description: Display nickname (3-24 chars, unique)
                  example: SpeedDemon
      responses:
        '200':
          description: Nickname updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  nickname:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /me/wallet/balance:
    get:
      tags: [Users]
      summary: Get wallet balance
      description: Fetches live balance from Arkeo chain
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Wallet balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: string
                    example: "1000000uarkeo"
        '503':
          description: Balance unavailable
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: racing_session
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: invalid_request
    User:
      type: object
      properties:
        userId:
          type: string
        handle:
          type: string
        nickname:
          type: string
          nullable: true
        walletAddress:
          type: string
          nullable: true
        balance:
          type: integer
    WalletInfo:
      type: object
      properties:
        walletAddress:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true
        chain:
          type: object
          properties:
            chainId:
              type: string
            chainName:
              type: string
            rpcUrl:
              type: string
            restUrl:
              type: string
            denom:
              type: string
            decimals:
              type: integer
            ticker:
              type: string
            bech32Prefix:
              type: string
...